--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -69,8 +69,8 @@ jobs:
               -DMACOS_BUNDLE=ON
               -DMODPLUG=ON
               -DWAVPACK=ON
-              -DVCPKG_TARGET_TRIPLET=x64-osx-min1012
-              -DVCPKG_DEFAULT_HOST_TRIPLET=x64-osx-min1012
+              -DVCPKG_TARGET_TRIPLET=x64-osx-min1012-release
+              -DVCPKG_DEFAULT_HOST_TRIPLET=x64-osx-min1012-release
             # TODO: Fix this broken test on macOS
             ctest_args: --exclude-regex DirectoryDAOTest.relocateDirectory
             cpack_generator: DragNDrop
@@ -92,8 +92,8 @@ jobs:
               -DMACOS_BUNDLE=ON
               -DMODPLUG=ON
               -DWAVPACK=ON
-              -DVCPKG_TARGET_TRIPLET=arm64-osx-min1100
-              -DVCPKG_DEFAULT_HOST_TRIPLET=x64-osx-min1012
+              -DVCPKG_TARGET_TRIPLET=arm64-osx-min1100-release
+              -DVCPKG_DEFAULT_HOST_TRIPLET=x64-osx-min1012-release
             # TODO: Fix this broken test on macOS
             crosscompile: true
             cpack_generator: DragNDrop
@@ -122,6 +122,8 @@ jobs:
               -DMEDIAFOUNDATION=ON
               -DMODPLUG=ON
               -DWAVPACK=ON
+              -DVCPKG_TARGET_TRIPLET=x64-windows-release
+              -DVCPKG_DEFAULT_HOST_TRIPLET=x64-windows-release
             cc: cl
             cxx: cl
             # TODO: Fix these broken tests on Windows