This updates to the latest vcpkg environment. 
This uses the new release prefix for the release only triplets like upstream.