@JoergAtGithub 2023-11-30
On Windows 11 the following 3 cases were tested successful:
x64_portable from release-only buildenv
x64_portable from full buildenv
x64_off from full buildenv

Due to the update of the libdjinterop VCPKG port to the required version, I got faster build times than ever:
Debug build (x64_off):  85s -> 66s
RelWithDebInfo: 115s -> 88s
(Full rebuild time of mixxx without mixxx-test)