@JoergAtGithub 2023-11-30
@fwcd or @m0dB 
Could one of you test this PR on Macintosh too.
Especially the case with the renamed triplet in the release-only buildenv archive. Does it still build for both supported CPU architectures?