@fwcd 2023-11-30
In that case, I'd say it's fine. While it would of course be nice to test the arm64-native build too, the x86-64 native and cross-builds from CI should be enough to catch anything major (for now).