@fwcd 2023-11-30
Ah sorry, for some reason I thought this was a PR in `vcpkg`, that makes sense. I'll have a look