@fwcd 2023-11-30
```
$ arch -x86_64 cmake -DCMAKE_TOOLCHAIN_FILE=/Users/<user>/git/mixxx/buildenv/mixxx-deps-rel-2.4-x64-osx-min1012-release-c82358c/scripts/buildsystems/vcpkg.cmake -B build
-- CMAKE_VERSION: 3.27.7
-- Using MIXXX_VCPKG_ROOT: /Users/<user>/git/mixxx/buildenv/mixxx-deps-rel-2.4-x64-osx-min1012-release-c82358c
-- Setting CMAKE_BUILD_TYPE to 'Debug' as none was specified.
-- The C compiler identification is AppleClang 15.0.0.15000040
-- The CXX compiler identification is AppleClang 15.0.0.15000040
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Detecting CXX compiler ABI info - done
-- Check for working CXX compiler: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Detecting CXX compile features - done
-- VCPKG_TARGET_TRIPLET dir not found: /Users/<user>/git/mixxx/buildenv/mixxx-deps-rel-2.4-x64-osx-min1012-release-c82358c/installed/x64-osx-min1012 Make sure the VCPKG build environment is installed and contains the build for the selected triplet.
CMake Error at CMakeLists.txt:59 (message):
  Did you download the Mixxx build environment using
  ´/Users/<user>/git/mixxx/tools/macos_buildenv.bat´
Call Stack (most recent call first):
  CMakeLists.txt:211 (FATAL_ERROR_MISSING_ENV)


-- Configuring incomplete, errors occurred!
```

We'll probably have to update CMakeLists to default to the right `VCPKG_TARGET_TRIPLET`