@fwcd 2023-11-30
This seems to be an issue that only happens when using `BUILDENV_RELEASE`