@fwcd 2023-11-30
I could see how this could become a bit annoying to check for in the CMakeLists, since release-mode dependencies do not necessarily imply building with `CMAKE_BUILD_TYPE != Debug` and we want to support all combinations of {release,debug} dependencies and {release,debug} Mixxx...