@fwcd 2023-11-30
Side note: `macos_release_buildenv.sh` is currently broken, since that script currently tries to execute `macos_buildenv.sh` instead of sourcing it.